Rename functions.py -> query_expressions.py
